# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading Terminal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get user watchlist
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add market to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /watchlist/{symbol}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove market from watchlist
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List user trades
      responses:
        "200":
          description: Trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Place a trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Invalid trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Trade not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    WatchlistItem:
      type: object
      required: [id, userId, symbol, addedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        symbol:
          type: string
        addedAt:
          type: string
          format: date-time
    WatchlistInput:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
    Trade:
      type: object
      required: [id, userId, symbol, side, quantity, price, status, mode, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        price:
          type: number
        total:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, filled, cancelled]
        mode:
          type: string
          enum: [paper, real]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        filledAt:
          type: ["string", "null"]
          format: date-time
    TradeInput:
      type: object
      required: [symbol, side, quantity]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        price:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...